+++ tests/main/test_config.py
@@ -8,7 +8,7 @@ from brownie._config import _load_config
 
 
 def test_base_keys(config):
-    assert set(config) == {"network", "pytest", "compiler", "colors", "active_network"}
+    assert set(config) == {"network", "pytest", "compiler", "active_network"}
 
 
 def test_network_keys(config):